xds: refactor to eliminate LookasideChannelLb
